01_data_transformation/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.14.1
google-cloud-core==2.3.1
google-cloud-pubsub==2.13.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
numpy==1.23.1
packaging==21.3
proto-plus==1.20.6
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
rsa==4.8
six==1.16.0
urllib3==1.26.10